requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.0.0
typing_extensions>=4.0.0

# LangChain stack (versões atualizadas para Python 3.14)
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# LangGraph
langgraph>=0.2.70
langgraph-checkpoint>=2.0.0

# OpenAI SDK
openai>=1.50.0

# Mercado Pago SDK (opcional)
mercadopago>=5.0.0

# Utilidades
requests>=2.30.0
tenacity>=8.0.0,<9.0.0  # Compatível com LangChain atual
tqdm>=4.60.0
PyYAML>=6.0.0

# Vetorização (alternativa ao ChromaDB, se precisar)
faiss-cpu>=1.8.0

# API e UI
fastapi>=0.100.0
uvicorn>=0.20.0
streamlit>=1.28.0

# Outros
httpx>=0.25.0
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.0
sqlalchemy>=2.0.0
pytrends>=4.9.0


# APIs de Redes Sociais
tweepy>=4.14.0  # Twitter/X
google-ads>=23.0.0  # Google Ads
linkedin-api>=2.0.0  # LinkedIn (opcional, ou usar requests)
# pinterest-api removido por incompatibilidade
# TikTok, Snapchat, YouTube usam requests direto

# Geração de PDFs
reportlab>=4.0.0
